The following adjustments are made: 1. Code Refactoring, 2. Integration with ActiveMQ broker, 3. Implementation of JPA and DTO validations, 4. Transaction management.

+++ codigo/bank/src/main/java/com/neoris/bank/dto/ReporteMovmientosDTO.java
@@ -1,8 +1,12 @@
 package com.neoris.bank.dto;
 
 
-import lombok.*;
-
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 
 @Setter
 @Getter
@@ -20,12 +24,12 @@ public class ReporteMovmientosDTO {
 
     private String tipo;
 
-    private int saldoInicial;
+    private long saldoInicial;
 
     private boolean estado;
 
-    private int movimiento;
+    private long movimiento;
 
-    private int saldoDisponible;
+    private long saldoDisponible;
 
 }
